refactor: merge utilities in one modules
This commit merges all modules under `utils/` in one module
`utils/index.js` and updates all javascript imports accross the
repository.

For organization sake, the `buildTree` in `utils/ops.js` was moved to
`views/browser/utils.js`. Since this function is only needed for
building the `Browser` component.

Everything else is simply formatting due to prettier and eslint.
