+++ frontend/src/views/browser/utils.js
@@ -1,24 +1,4 @@
-const groupBySplitME = (data) => {
-  const groupedData = data.reduce((acc, item) => {
-    const [firstPart, secondPart] = item.me.split('/')
-    const key = `${firstPart}/${secondPart.split('/')[0]}`
-
-    if (!acc[key]) {
-      acc[key] = {
-        me: key,
-        count: item.count,
-      }
-    } else {
-      acc[key].count += item.count
-    }
-
-    return acc
-  }, {})
-
-  return Object.values(groupedData)
-}
-
-const buildTree = (items) => {
+export const buildTree = (items) => {
   const tree = []
   items.forEach((item) => {
     const path = item.me
@@ -42,5 +22,3 @@ const buildTree = (items) => {
   })
   return tree
 }
-
-export { groupBySplitME, buildTree }
