+++ frontend/src/views/dataExplorer/runs.jsx
@@ -9,11 +9,7 @@ import Row from 'react-bootstrap/Row'
 import { toast } from 'react-toastify'
 
 import API from '../../services/api'
-import {
-  isNumericNonZero,
-  isStringNonEmpty,
-  getNextToken,
-} from '../../utils/sanitizer'
+import { isNumericNonZero, isStringNonEmpty, getNextToken } from '../../utils'
 import Table from '../components/table'
 
 const Runs = () => {
